always there...

+++ delete_network.php
@@ -2,19 +2,23 @@
 	include('functions.php');
 
 	$id = "";
-	if(array_key_exists("id", $_GET)) {
-		$id = $_GET["id"];
-		if(can_edit_models(get_user_id_from_model_id($id))) {
-			if($id == "") {
-				print "This network doesn't exist.";
+	if(is_admin()) {
+		if(array_key_exists("id", $_GET)) {
+			$id = $_GET["id"];
+			if(can_edit_models(get_user_id_from_model_id($id))) {
+				if($id == "") {
+					print "This network doesn't exist.";
+				} else {
+					$user_id = get_user_id_from_session_id($_COOKIE["session_id"]);
+					delete_model($id, $user_id);
+				}
 			} else {
-				$user_id = get_user_id_from_session_id($_COOKIE["session_id"]);
-				delete_model($id, $user_id);
+				print "You don't have the permission to delete.";
 			}
 		} else {
-			print "You don't have the permission to delete.";
+			print "Network name doesn't exist.";
 		}
 	} else {
-		print "Network name doesn't exist.";
+		print "Only admins cat set models to public";
 	}
 ?>
